Reworked to force override.
